run ALL tests locally (#51)
## Context

We should be able to run locally all the tests that we run in the CI to
not overload the CI itself that has a limited amount of minutes per
month.

## Scope

#### tox command changelogs
- [x] remove `bbp-model` command: it was referring to bbp, too
encompassing, not clear what was tested
- [x] added: `test-integration-e2e, test-scientific,
test-scientific-ngv` to test the respective test sets
- [x] ~~renamed: `test-unit, test-baseline, test-integration` so that it
is clear what it is doing. It is also the same convention as in the
folder naming (all the tests are under `tests/*`)~~ EDIT: @WeinaJi was
opposed to this. This change has been reverted
- [x] added: `tests` to run all the tests except `ngv`
- [x] added: `clean` to remove all the leftovers of the tests that we
run. To be used only manually, after manually checking the test results
- [x] updated: gitignore so that we do not accidentally commit leftovers
from the tests

#### workaround and fixes
Workaround because of #3
fix: #64 (thx @WeinaJi  for this collaborative effort)

## Testing

Atm, with these changes, all the tests pass in my local system.
Remember to update your `neurodamus-models` to 2.4.2. Otherwise
neurodamus will fail with a segfault in the `scientific` tests. It is
due to the previous pr about hoc maps