test stimuli (#69)
## Context
We need to expand our testing campaign.

Fix: #60 #62 #77 #78

## Scope
Stimuli is a well defined chunk of code that can be easily tested
extensively. Atm we have 2 test files that tests only parts of the
stimuli object, one in unit and one in integration tests. They should be
collected in 1 file that fully tests everything in the stimuli file.

It should be done as an integration test because we need the full
h.Vector. The mocked version of that used in the unit tests is not
enough for the alrady existing integration tests. The h.Vector is
fundamental for the stimuli object because of the `play` capabilities.

#### Tests
- [x] `SignalSource` methods
  - [x] remove `while ev[-1] < duration:` as we always overshoot
- [x] emit error for shot pulse with `tau_D == tau_R` since we do not
know what should the function be. Better than a divide by 0 error.
- [x] `CurrentSource` methods
- [x] ~~remove `stim_vec_mode` since it is always set to `True`. Let me
know if `_Clamp` is used explicitly somewhere~~ EDIT: used by `Constant`
- [x] `ConductanceSource`
- [x] remove `stim_vec_mode` as it is a copy-paste from `CurrentSource`.
However, since `Constant` is missing here, it is always True
  - [x] fix behavior around 0 for `ConductanceSource`

- [x] after discussion, add assert to be sure that we do not have
regressions

## Testing
This is TESTING!!! (it should sound as: this is SPARTAAA!)

## Blockers:
- [x] Decide what to do with: #77 

## Review
* [ ] PR description is complete
* [ ] Coding style (imports, function length, New functions, classes or
files) are good
* [ ] Unit/Scientific test added
* [ ] Updated Readme, in-code, developer documentation