Unit tests for loading 2 populations and check cell states in NEURON context (#71)
## Context
As required in #59 , this PR adds unit tests which loads 2 populations
from the ring circuit for NEURON or CORENEURON simulator. The tests also
check the cell/synapses/netcons states in the NEURON context.

## Scope
In test_ringcells.py:
- improve the existing check function `_check_synapse` and
`_check_netcons` that can be shared across tests
- improve the test `test_dump_RingA_RingB` to check
cell/synapses/netcons in addition to the check of dump files
- tests use simulator "NEURON" and "CORENEURON"

## Review
* [x] PR description is complete
* [x] Coding style (imports, function length, New functions, classes or
files) are good
* [x] Unit/Scientific test added
* [ ] Updated Readme, in-code, developer documentation