`spont_minis` simple test (#124)
## Context

`spont_minis` uses a stochastic process to generate events. Synapses, if
ready, release generating a voltage change.

## Scope
Proper testing of this requires a stochastic test campaign that is out
of the scope of this PR.

Here I just want to see that I see a voltage change in `RingB` if there
is a `spont_minis` attached. Without the `spont_minis` no peaks are
recorded. Thus, checking the precise position of the peaks for this
`spont_minis` should be sufficient as reproducing this randomly or from
other sources is very unlikely.

I am also testing that a pont minis is attached to the synapse and that
in the weight it really serves its type as spont_minis

- [x] remove accidental debug code that remained around

## Fix

#114 only partially,


## Additional bug fixes

#123 while working on this we noticed that `sim_init` can be called
twice (it has protection) but `init` does not. I added the same
protection to `init` too.

## Review
* [ ] PR description is complete
* [ ] Coding style (imports, function length, New functions, classes or
files) are good
* [ ] Unit/Scientific test added
* [ ] Updated Readme, in-code, developer documentation