synapse bluecellulab checks (#144)
## Context

After talking with bluecellulab we decided to be inspired by some of
their tests (copy). This PR adds some checks to the synapses and the
ways you can override values with the `simulation_config.json` that I
was not aware of.

## Fix #142 

## Scope

- [x] `nrrp` was already checked. Minor improvements: it should be a
positive integer
- [x] `"%s.NMDA_ratio = 1.22 %s.tau_r_NMDA = 3.9 %s.tau_d_NMDA =
148.5"`.
    - [x] `NMDA_ratio`
- [x] `tau_r_NMDA` and `tau_d_NMDA` are global variables. Added test to
override these
- [x] add test to verify synapse location
- [x] test in particular modoverride for `GluSynapse`
- [x] move/merge `test_connection.py` into unit tests
- [x] check synapse override from
[conditions](https://sonata-extension.readthedocs.io/en/latest/sonata_simulation.html#parameters-required-for-modifications)
and how it interacts with the one from
`connection_overrides.synapse_configure`. Spoiler:
``connection_overrides.synapse_configure` wins

## Additional improvements, not stemming from bluecellulab

- [x] `np.isclose(10001.1, 10001.2) == True`. We should use smaller
values. default rtol is 1e-5
- [x] modoverride test was not really overriding (it was already ampa).
Switch to gaba in the test
- [x] fix pytest assert reporting if `tests.utils` is imported in
`conftest.py`
- [x] create issue for inconsistend behavior for synapse global variable
override #147
- [x] fix
[docs](https://github.com/openbraininstitute/sonata-extension/pull/2)