`Node.__init__`: split it into `__init__` and `_reset`
**State of the Art**

`__init__` has an internal api where, if the config_file is None
it just resets the internal variables. It is done mainly for
multicycle runs. This goes against the SRP. In addition, users
may still use this function in this way. This may create confusion.

**What was done**

`__init__` has been split in `__init__` and `_reset`. `__init__`
calls reset under the hood. The main advantages are that:

- `_reset` is clearly marked for internal use only
- `__init__` simply fails if the `config_file` is `None`
- SRP

Added also a nice docstrings for good measure

**Testing**

Since this is refactoring and `__init__` is already tested this
should not alter our tests and no additional tests should be required.