progressBar tests (#165)
## Context

ProgressBar was untested. Let's move some tests to the unit tests
folder.

This stems from a very evasive bug that I encountered while checking
other stuff. I cannot find a way to reproduce the problem in a unit
test. I leave the solution but I do not know how to test it.

I need to close stdout but all my efforts are in vain

## Scope

Progressbar unit tests


## Review
* [ ] PR description is complete
* [ ] Coding style (imports, function length, New functions, classes or
files) are good
* [ ] Unit/Scientific test added
* [ ] Updated Readme, in-code, developer documentation