`ngv` docs (#159)
## Context
NGV has been left without documentation and with a lot of dead ends. We
need a thorrough review and a rework of the whole system.

I wanted to go much deeper, prepare a new circuit and all but I do not
have time as we need to test coreneuron. However, these docs can go to
main