Remove code for forwardskip, deprecated in SONATA (#161)
## Context
ForwardSkip was used in legacy BlueConfig file. In SONATA config files,
we've decided to deprecate this parameter as what has been done in
libsonata parser. Therefore, this PR removes the relevant code for
forward skip in NEURON simulator. It can not be set in SONATA config file.

## Scope
Remove both the forward skip in NEURON and in CoreNEURON configurations.

## Testing
no testing as it is not used in SONATA.

## Review
* [x] PR description is complete
* [x] Coding style (imports, function length, New functions, classes or
files) are good
* [ ] Unit/Scientific test added
* [ ] Updated Readme, in-code, developer documentation