multicycle coreneuron test (#166)
## Context

CoreNeuron has the multicycle functionality. It is already partially
tested in e2e. We should add a test in unit too. Check that the results
are the same as for the filemode test.

The important part of the code may be
[here](https://github.com/openbraininstitute/neurodamus/blob/9066f6beff0a89a14bafd0137fb1a426d4f49827/neurodamus/node.py#L1911)

## Scope

- [x] Rename `test_coreneuron` -> `test_coreneuron_filemode` to clarify
its purpose
- [x] Add new `test_coreneuron_multicycle` test file 
- [x] Reorganization of the existing e2e multicycle tests. 1 can go in
the unit tests for the nodeSets. The others remain in e2e because some
mechanisms are not installed for the unit tests.
- [x] improve docstrings
- [x] limit the max number of multicycle steps to the max number of
nodes per population. This is in line to the nodeset distribution
algorithm used down the line
- [x] add an assert to be sure that when we add a fake gid for avoiding
coreneuron complains we are not adding into an existing population
- [x] `isFile` -> `is_file` (typo)
- [x] split `_instantiate_simulation` in a few functions. Why? SRP 

Fix: #153, #176

## Testing

This is testing