save/restore (#177)
## Context

Neurodamus offers save/restore capabilities. We should test them. More
in #154

## Scope

`Node, CoreConfig and SimConfig` have been reworked so that when we call
a save in a folder we really save all that is needed for a restore in
that folder. If no save is provided, we put the files in output. If no
`--keep-build`, some of these files are then removed.

Test neurodamus save/restore features on ringtest on various cases for
both neuron and coreneuron.

- We repeat identical coreneuron simulations where we dump the cell
state of a gid. Unfortunately we cannot get the cell state of more than
a gid at a time (coreneuron limitation).
- We collect the results and check that the cell states if we run
immediately at t[2] is equal to the cell state if we do a save restore
at t[1]

## Note

After trying for a while to change all the pathing in SimConfig to make
them as properties from classmethods I reached the conclusion that it is
not trivial. Here is why:

`@classmethod` and `@property` cannot be combined in 3.13. `@property`
can be used only on instances. If a path is used in a `@classmethod` it
does not have access to the instance (only the class). Thus, some paths
cannot changed to `@property` easily

## Detailed Changelog

### Major

- [x] remove obsolete `integration-e2e/test_cli_opts.py` tests referring
to save/restore
- [x] remove obsolete `save_time` cli option as coreneuron cannot use it
as neuron cannot use the save function anymore
- ~~add multi-rank tests~~ moved to #193
- [x] splitting of some paths in `*_save` and `*_restore` to
differentiate if they are to read or to written. They need to remain
classmethods in SimConfig
- [x] `populations_offset` is copied because it is needed for a restore
as is and needed in output
- [x] `coreneuron_input` cannot be split in save and restore: coreneuron
limitation. Thus, it is linked into the save folder from the restore
folder. It was already like this but in a random place of the code. Now
everything is in `Node._setup_coreneuron_restore_environment`
- [x] `CoreConfig` does not hold data of its own. It spies inside
`SimConfig` to get the paths. In this way we do not have sync problems
anymore. 1 source of truth. This implies that `SimConfig` cannot `import
CoreConfig`. We can have them as properties because no classmethod uses
these properties
- [x] improved `update_tstop` so that the file it is not written
`n_reports` times. Since the file contains binary file it may be big.
Rewriting multiple times was really a waste.
- [x] write docs about what paths should contain
- [x] new issue: create a test of the shared memory test (shm) for mac
#209
- [x] new issue: have coreneuron write 1 file instead of 1B in
`coreneuron_input` #210
- [x] incorporate #204
- [x] manual test for mpi save/restore (fixed Jim's bug) 
- [x] add deprecation issue for `_create_save_handler` #212

### Minor

- [x] remove `__slots__` from SimConfig. In this way the class members
are not read-only. slots saves noticeable amounts of memory only if many
instances of the class are required
- [x] standardize `test_ringcells`
- [x] adjusted some tests to address the new paths in SimConfig and
CoreConfig

## Issues

Fix: #154, #178

## Testing

This is testing

Co-authored-by: Weina Ji <weina.ji@openbraininstitute.org>