improved `dump-cell-state` (for neuron only, old behavior for coreneuron) (#218)
## Context
Now that I have a nice base with the save/restore tests I can proceed to
improve/test `dump-cell-state`. Currenty `--dump-cell-state=<gid>`
accepts only 1 gid at a time. If neuron is used, there an undocumented
way to dump more than 1 gid at once: you need a file called
`debug_gids.txt` (the name is hardcoded) in the base simulation folder.
In that case neurodamus reads the gids from that file and dumps
everything. This is done by `dump_circuit_config`.

This can be largely improved (at least for neuron)

Fix: #220

## Scope

In this PR I want to be able to write:
`--dump-cell-state=0,1,2,3-9,8,7,10-13` and dump these gids with neuron.
Unfortunately, this functionality cannot be put in place for coreneuron
as it requires either:

- changing coreneuron
- having a functioning parallel context, not always true when using
coreNeuron

In case coreNeuron is used, we rever to the old behavior and use only
the first gid (emitting a warning).

- [x] update `--help`
- [x] rename a few variables to have them more consistent 

## Testing
- [x] add unit tests