mpi tests (proof of concept) (#213)
## Context
We want to test neurodamus also with mpi #194. However, this turns out
to be trickier than expected. There are multiple libraries around with
pros and cons. The following list is not exaustive; it is just what I
have found so far and my opinion after bashing my head on all these
options:

- **pytest-mpi**: it is very non-intrusive. It just makes pytest aware
that is run in mpi. The standard way to run it is `mpiexec -n 2 pytest
--with-mpi ...`. The problem is that this interpheres with `--forked`.
Either the processes are not forked or the mpi fails. In addition, it is
ugly to see messages from various ranks interphere with each other. I
did not investigate further
- **mpi-pytest**: different from pytest-mpi, newer, it makes pytest call
`mpiexec` on each test that requires that. A typical call is `pytest
tests/unit`. Inside `mpi-pytest` calls `mpiexec`. Unfortunately it works
only with MPICH as we easily get nested `MPI_Init` calls. Changing the
code to address this problem is too much IMHO
- **HBP_STEPS**: the simulator is more different than expected from our
case. Over there we use `ctest` to discover the tests and we use
`unittest` to test. Our problems are effectively avoided but we cannot
easily replicate the same behavior in out setup.
- **pytest-isolate-mpi**: test suite to run programs with mpi in
isolated environments. Ufortunately, if we do not run with mpi, the
suite does not run the tests forked and the 2 libraries (`pytest-forked`
and `pytest-isolate-mpi` interphere with each other when neuron and its
mechanisms are involved). (chosen solution)
- **in house**: creating our own test suite is a burden and becomes also
complex to maintain. We can tailor it exactly to our needs

## Scope

I picked `pytest-isolate-mpi` as it provides already a lot of what we
want. The problem is that it requires a slightly different setup than
what we have in the unit tests (for example).

- [x] add a new tox environment where we can require
`pytest-isolate-mpi`
- [x] add tests to check standard MPI capabilities
- [x] add tests to check MPI + neurodamus capabilities (check also that
the processes are forked)
- [x] move `ringtest_baseconfig` to base `conftest.py`
- ~~check if and how `--basetemp` works~~ EDIT: `basetemp` is ignored
when `pytest-isolate-mpi` is employed. We can still set `TMPDIR` in
`setenv` in `tox.ini`
- ~~make it work out of the box even with our base setup (unit, etc)~~
EDIT: `pytest-isolate-mpi` discards `--forked` when no mpi command is
employed. It is likely a bug. The workaround is to mark all the
functions with `@pytest.mark.mpi(ranks=1)`. However, this is
labour-intensive. Probably having 2 setups is simpler
- [x] coverage
- [x] merge #222 from @WeinaJi 
- [x] add test with empty rank (1 cell, 2 ranks) 

## Testing

Added some sample tests that check that mpi processes are properly
forked and repeatible. Saving files to possibly catch race conditions
too

---------

Co-authored-by: Weina Ji <weina.ji@openbraininstitute.org>