Cleanup layers of indirection in sonata_config.py::SonataConfig (#133)
For historical reasons, the SONATA config was being loaded and
translated heavily.

The in this PR include:
* remove the `def __getattr__(self, item)`: this made things difficult
to find and debug.
for instance, if one references `self.network`; this was doing a
"translation", and was being returned from
one of the `_entries`, `_sections`, or, after some manipulation, a
`_translation`
* reduce redundant copies of the same information, less likely to modify
one, and then another being used:
ie: removed the fields, since they weren't necessary, and could be
accessed from the others:
_circuit_networks, _config_dir, _config_json, _entries,
_resolved_manifest, _sections,
* removed custom implementation of functionality that exists in
libsonata; this fixes defaults that were being incorrectly handled (see
#217)
* fixed tests that were relying on `v_init` being -65, even though the
default should be -80. in SONATA:
seehttps://github.com/openbraininstitute/neurodamus/issues/217