astrocytes: tests + refactoring (#252)
## Context
Astrocytes may be tested better. The implementation can be largely
improved. There will be additional prs. I am breaking the refactoring in
manageable pieces so reviewing them is easier.

## Scope
### Astrocyte
- [x] incorporate `_init_cell` in `__init__`
- [x] removed `glut_list` are a real member variable. We use now
`sections_glut` and `soma_glut` variables. `glut_list` is a `@property`
for compatibility with metabolism. This simplifies and makes the code
more robust against off by 1 errors.
- [x] fix #251
- [x] ~~remove hoc from `create_endfeet`~~ EDIT: moved to #258
- [x] add issue for `_nseg_warnings` #261
### MorphIOwrapper
- [x] remove `section_index2name_dict` from morphIOwrapper. replaced
with `section_names`
- [x] fix #148  
- [x] add a new `@dataclass` to handle section names: `SectionName`. It
has also a very handful `__str__` member function to auto-convert it in
standard neuron/hoc code


## Bugfixes
### Astrocyte
- [x] `GlutReceiveSoma` should be just used for reporting. It should not
interact with `cadifus`. Atm we effectively override `GLutReceive` in
the soma section with this one for what concerns `cadifus`. We stop with
that. fix #257

## Testing
### Astrocyte
- [x] move astrocyte testing to new file. GlutList included
- [x] add testing for `GlutList`
- [x] change reference in old ngv tests since they were generated with a
bug in the code
- [x] rearrange reference in tests as numpy expects the reference as
last (second) element of a comparison... Otherwise it fails pointing at
inverted ACTUAL and CURRENT values
- [x] change test so that there is 1 astrocyte with 2 endfeet
- [x] full comparison all Rad and R0pas in old radii test 
## MorphIOwrapper
- [x] test `get_section_names()`. Compare it with printnames? check the
types