ngv refactoring + tests (#269)
## Context
ngv can be refactored and tested more

## Scope
- [x] rework `_connect_endfeet` and dump the construction and endfeet
connection to the astrocyte in the astrocyte class. It belongs there.
Fix #259
- [x] remove endoplasmic reticulum functions. The functions were unused
- [x] add in the docs what we were doing with the endoplasmic reticulum.
Partial fix #267
- [x] add `_init_section` and `_init_endfoot_section`: it is more SMART
- [x] fix #261: better warning for section resizing
- [x] renamed `*glut` lists in `glut_all`, `glut_endfeet`, and
`glut_soma` to underline where they are from. `glut_list` remains for
compatibility with metabolism
- [x] simplified `R0pas` extraction. I do not think an endfoot without a
vasculature has much sense
- [x] fix #262
- [x] fix #263 not necessary anymore. From Michael Hines 
>I do believe all ref_vars are automatically updated.
CVode.cache_efficient now does nothing. In version 9 all range variable
memory layout is Structure of Array. The substance of cache efficiency
is now merely a permutation.
Replaced with the more pythonic: `sec(0.5).cadifus._ref_glu2 =
glut._ref_glut`.
- [x] fix #276. No answer in the ticket. I decide: we can emit a normal
warning when we encounter the problem and avoid the whole `post_stdinit`
function which, in turn, can be safely removed now.

There is a test to check that it still works as a pointer
- [x] fix #272 make vasculature mandatory
- [x] `GlioVascularManager` should use proper `ConfigurationError`, not
`Exception`
- [x] skip refactoring of `NeuroGlialConnection` as it requires
unblocking of #149 and this fix is likely to be a PR by itself based on
its complexity
- [x] fix #269
- [x] fix #149: we remove the semi-implemented the feature
- [x] fix #285: `GlutReceive` is placed only when needed. `get_glut`
ensures that we place, if needed, at most 1 `GlutReceive` per section
- [x] added a new `_emit_init_cell_warnings` that collects and emits
warnings from cell inits in rank 0. Testing this is going to be
complicated because I need to create a new cell type, a manager that
uses that and then I need to somehow init it. I think to init that I
need a config object and a population which requires various files and
configuration data. Maybe it is just an overkill just to see if this
function correctly reports the warnings. @WeinaJi what is your opinion?
- [x] remove `netcon.record` debug messages

## Testing
- [x] add real unit tests for the `Astrocyte`