CI: build and test simulation toolchain with macOS-15 (#330)
## Context
Add `macOS-15` in `simulation_test.yml` to test the build and run unit
tests.

## Scope
- `macOS-15` with python 3.12.6. We could test more python versions in
macOS but there is the cache limit in GH actions, so choose one python
version atm.
- remove the action cache for python virtual env. The cache is built for
the python version and the libsonata tag used in the pipeline, but it
won't rebuild if any python package in the venv is updated. And `pip
install` is fast to build all required packages so we don't gain much
(1-2 mins) with cache.
- remove `macOS` test in `coverage.yml`

## Review
* [x] PR description is complete
* [ ] Coding style (imports, function length, New functions, classes or
files) are good
* [ ] Unit/Scientific test added
* [ ] Updated Readme, in-code, developer documentation