report improvements (#338)
## Context
Simplify reports, enhance and improve compartment and summation reports
to report on many more different variables.

Since I had to do this anyway because:
- compartment and compartment sets share a lot of routines
- it is not clear in the documentation if something is restricted or it
is just a bug. I clarified and expanded the capabilities of the code
wherever it was possible

All of this to give me a clear overview on how to implement the
compartment sets reports in coreneuron

## Scope
- [x] expand compartment and summation reports to report also report
distributed mechanism values
- [x] unify variable parsing
- [x] unify scaling factors (only for summation reports)
- [x] remove unused parameter from `setup_alu_for_summation`

## Testing
- [x] added a test for the pas mechanism in `test_ringcells_reports.py`
- [x] add tests in `integration-e2e/test_reports.py`