compartment sets (#308)
## Context

- fix #307
- https://sonata-extension.readthedocs.io/en/latest/sonata_report.html#fine-grained-compartment-report

In addition to this, while I was touching the reports, I proceeded to:

Simplify reports, enhance and improve compartment and summation reports
to report on many more different variables.

Since I had to do this anyway because:
- compartment and compartment sets share a lot of routines
- it is not clear in the documentation if something is restricted or it
is just a bug. I clarified and expanded the capabilities of the code
wherever it was possible

All of this to give me a clear overview on how to implement the
compartment sets reports in coreneuron

### Additional Fixes

- fix #309
- fix #310

## Scope
- [x] add `compartment_sets_file` to parsed_run. Now it may be necessary
- [x] `section_index` -> `section_id`: homogeneization of names to
dissipate confusion. Libsonata spec talks about `section_id`.
- [x] have proper section_id support to move interchangibly
(section_type, id) <-> section_id. Placed in BaseCell for now
- [x] make `_get_sec` in `ngv` use the new routine in metype and remove
the unnecessary map from IOwrapper
- [x] `section_names` -> `_section_names` in morphIOwrapper
- [x] move the section_index tests in tests for `metype.py` + add test
for `get_sec`
- ~~coreneuron reports~~: moved to #322
- [x] add error for coreneuron reporting (not implemented yet)
- [x] `get_section_id` raises an error if we ask for an id of a section
that is not present (relies on `nSec*`)
- [x] `get_sec`: `section_id` is now assumed to be based on the original
cell (with the axon). Thus, this is the expected
behavior:
   - if we ask for a sec in axon that was removed -> raise error
- if we ask for a sec in axon that is still there (section 0 or 1) ->
return that
- if we ask for a sec "after" the axon (any section type that is not
soma) we compute the offset based on the orignal cell (with the axon)
- [x] do not rely on `nSec*`

Report-related improvements:
- [x] expand compartment and summation reports to report also report
distributed mechanism values
- [x] unify variable parsing
- [x] unify scaling factors (only for summation reports)
- [x] remove unused parameter from `setup_alu_for_summation`

## Testing
- [x] added test in `integration-e2e/test_reports.py`
- [x] add test to compare full report vs compartment set
- [x] test `get_section_id` and `test_get_sec`

Report-related testing:
- [x] added a test for the pas mechanism in `test_ringcells_reports.py`
- [x] add tests in `integration-e2e/test_reports.py`