compartment_sets: additional test
I noticed that I never tested compartment sets when multiple
populations are involved. Ringtest is the best suited
because there are already 2 pops there.

Additional changes:

- add debugging code

format

additional test

- add test for compartment set report without cached targets. It may
fail in the future. This is a good test in case libsonata changes
and it does not provide the default cells: Mosaic for a report

- add reference file for one more test

fix docstring

WIP