enable `cell_permute` for coreneuron (#375)
## Context
`corenrn` can run with `cell-permute = {0 (cpu, gpu), 1 (cpu, gpu), 2
(gpu)}`. Neurodamus at the moment is not leveraging this. We should
allow users to set this.

**Note:** this is different from `cell-permute` for neuron (9+ only). I
may enable that in the future. In that case this will become useless
legacy code. At any rate, neuron 9 is new and not fully used everywhere.
Proceeding step by step, enabling `cell-permute` in coreneuron is the
next logical step

## Scope
* add option in CLI and in the python bindings

## Testing

* test that it is written in `sim.conf`
* test that `integration-e2e/test_reports.py` is unaffected