fix test_conn_manager_syn_stats
This test was using 0-based targets while get_conn_counts expects
1-based input. This, in turn, made the 0 gid to be turned in
a 4 billion gid (out of the local nodes). It was still counted
in the stats as a gid without synapses.

The edge file says that there are 2 synapses from 0 to 1 and 2 from 1
to 0. So:

- gid 0 was going to inf. No synases but it is a counting gid with metype x
- gid 1 was going to 0. It has 2 synapses. On average metype x has 1 because of
gid 0 (now 4 billion) not having synapses.
- gid 2 was going to 1. metype y has 2 synapses (correct).